Minor cleanup and rework build script

+++ build.sh
@@ -1,9 +1,27 @@
 #!/bin/bash
 
-meson setup build
-meson compile -C build
+CLEAN_BUILD=${1:-"null"}
+BUILD_OUTPUT=""
 
-# Clean up build
-rm build/game.h build/game.vapi
+if [ "$CLEAN_BUILD" = "clean" ]; then
+    if [ -d "build" ]; then
+        rm -r build
+    fi
 
-build/virgil --directory=build --show-fps
+    BUILD_OUTPUT=$(meson setup build)
+
+    if [ $? != 0 ]; then
+        echo "$BUILD_OUTPUT"
+    fi
+fi
+
+BUILD_OUTPUT=$(meson compile -C build)
+
+if [ $? = 0 ]; then
+    # Clean up build
+    rm build/game.h build/game.vapi
+
+    build/virgil --directory=build --show-fps
+else
+    echo "$BUILD_OUTPUT"
+fi
